requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
django-cors-headers==3.8.0
django-jazzmin==2.4.8
django-summernote==0.8.11.6
djangorestframework==3.12.4
django-modeltranslation==0.17.1
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.3.1
python-decouple==3.4
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6